use competing factor to reduce the number of necessary samples.
we can lower the threshold to just 0.85,
as we have another measure to capture competitors:
if there are multiple values that are better
than the threshold but tagged differently,
we will get a "competing factor".
but so far we haven't got any competing factor,
meaning the threshold alone is good enough for the time being.